Update patch set 4
Patch Set 4:

> On a unrelated note (have you got jenkins working with zuul v3?)

There's no place for Jenkins in Zuul v3, and there was only limited place for Jenkins in Zuul v2. The only task which was still being done by Jenkins was "just launching the jobs somewhere". Internally, we have never used Jenkins for that, only turbo-hipster on v2. We are (slowly) migrating to v3's native Ansible bits with nodepool on v3.

So, nope -- it doesn't make sense to use Jenkins with Zuul v3. Convert your jobs to Ansible, or to shell scripts if you would like to.

Patch-set: 4